Code clean-up Updated to target SDK 34

+++ app/src/main/java/net/vertexgraphics/myfinances/LogActivity.java
@@ -2,20 +2,22 @@ package net.vertexgraphics.myfinances;
 
 
 import android.app.Dialog;
-import android.content.DialogInterface;
 import android.os.Bundle;
-import androidx.core.app.NavUtils;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import android.text.method.ScrollingMovementMethod;
 import android.view.Gravity;
-import android.view.KeyEvent;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.app.NavUtils;
+
+import java.util.Objects;
 
 public class LogActivity extends AppCompatActivity implements View.OnClickListener
 {
@@ -24,7 +26,7 @@ public class LogActivity extends AppCompatActivity implements View.OnClickListen
 	TextView logText;
 	String[] logContentLines;
 	Toolbar toolbar;
-	private CustomSoftKeyboard keyboard;
+	//private CustomSoftKeyboard keyboard;
 	
 	@Override
 	protected void onCreate(Bundle savedInstanceState)
@@ -32,15 +34,15 @@ public class LogActivity extends AppCompatActivity implements View.OnClickListen
 		
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.log);
-		toolbar = (Toolbar) findViewById(R.id.toolbar);
+		toolbar = findViewById(R.id.toolbar);
 		
 		toolbar.setTitle(getString(R.string.log_string));
 		setSupportActionBar(toolbar);
-		getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+		Objects.requireNonNull(getSupportActionBar()).setDisplayHomeAsUpEnabled(true);
 		
 		
 		userPrefs = UserPrefs.getInstance(this);
-		logText = (TextView) findViewById(R.id.logText);
+		logText = findViewById(R.id.logText);
 		logText.setMovementMethod(new ScrollingMovementMethod());
 		
 		
@@ -63,67 +65,38 @@ public class LogActivity extends AppCompatActivity implements View.OnClickListen
 		TextView label;
 		Button accept;
 		Button cancel;
-		View.OnClickListener cancelListener = new View.OnClickListener(){
+		View.OnClickListener cancelListener = p1 -> dialog.dismiss();
 
-			@Override
-			public void onClick(View p1)
-			{
-				dialog.dismiss();
-			}
-			
-			
-		};
-		
-		switch (item.getItemId()){
-			case R.id.filter:
-				
-				dialog.setContentView(R.layout.edit_name_dialog);
-				editText = (EditText) dialog.findViewById(R.id.editnamedialogEditText);
-				dialog.getWindow().setGravity(Gravity.BOTTOM);
-				keyboard = new CustomSoftKeyboard(dialog, R.id.keyboard, R.layout.qwerty_keboard, false, R.id.editnamedialogEditText);
-				keyboard.registerEditText(editText);
-				label = (TextView) dialog.findViewById(R.id.editnamedialogLabel);
-				label.setText(getString(R.string.setFilter_string));
-				cancel = (Button) dialog.findViewById(R.id.editnamedialogButton2);
-				cancel.setOnClickListener(cancelListener);
-				accept = (Button) dialog.findViewById(R.id.editnamedialogButton1);
-				accept.setOnClickListener(new View.OnClickListener() {
+        int itemId = item.getItemId();
+        if (itemId == R.id.filter) {
+            dialog.setContentView(R.layout.edit_name_dialog);
+            editText = dialog.findViewById(R.id.editnamedialogEditText);
+            Objects.requireNonNull(dialog.getWindow()).setGravity(Gravity.BOTTOM);
 
-						@Override
-						public void onClick(View p1)
-						{
-							updateLog(editText.getText().toString());
-							
-							dialog.dismiss();
-						}
-						
-					
-				});
-                dialog.setOnKeyListener(new DialogInterface.OnKeyListener() {
-                    @Override
-                    public boolean onKey(DialogInterface dialog, int keyCode, KeyEvent event) {
-                        if (keyCode == KeyEvent.KEYCODE_BACK) {
-                            keyboard.hideCustomKeyboard();
-                        }
-                        return true;
-                    }
-                });
-				dialog.show();
-				break;
-				
-				
-			case R.id.item_deleteLog:
-				userPrefs.deleteLog();
-				updateLog("");
-				break;
-				
-			case android.R.id.home:
-				NavUtils.navigateUpFromSameTask(this);
-				return true;
-				
-			
-				
-		}
+            label = dialog.findViewById(R.id.editnamedialogLabel);
+            label.setText(getString(R.string.setFilter_string));
+            cancel = dialog.findViewById(R.id.editnamedialogButton2);
+            cancel.setOnClickListener(cancelListener);
+            accept = dialog.findViewById(R.id.editnamedialogButton1);
+            accept.setOnClickListener(p1 -> {
+                updateLog(editText.getText().toString());
+
+                dialog.dismiss();
+            });
+
+            dialog.setOnShowListener(p1 -> {
+                editText.requestFocus();
+                dialog.getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_ALWAYS_VISIBLE);
+            });
+
+            dialog.show();
+        } else if (itemId == R.id.item_deleteLog) {
+            userPrefs.deleteLog();
+            updateLog("");
+        } else if (itemId == android.R.id.home) {
+            NavUtils.navigateUpFromSameTask(this);
+            return true;
+        }
 		
 		
 		
@@ -151,23 +124,15 @@ public class LogActivity extends AppCompatActivity implements View.OnClickListen
 		logText.setText("");
 
 		for(String s: logContentLines){
-			if(!s.equals("")){
+			if(!s.isEmpty()){
 				logText.append(s+ "\n");
 			}
 		}
 	}
-	@Override
-	public void onClick(View p1)
-	{
-		// TODO: Implement this method
-	}
+
 
 	@Override
-	public void onBackPressed() {
-		if(keyboard!= null && keyboard.isCustomKeyboardVisible()){
-			keyboard.hideCustomKeyboard();
-		}else {
-			super.onBackPressed();
-		}
+	public void onClick(View view) {
+		//TODO
 	}
 }
